move Resource to fiat-api
