+++ fiat-roles/src/main/java/com/netflix/spinnaker/fiat/providers/DefaultResourcePermissionProvider.java
@@ -17,7 +17,7 @@
 package com.netflix.spinnaker.fiat.providers;
 
 import com.netflix.spinnaker.fiat.api.Permissions;
-import com.netflix.spinnaker.fiat.model.resources.Resource;
+import com.netflix.spinnaker.fiat.api.Resource;
 import java.util.Objects;
 import javax.annotation.Nonnull;
